+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 * Method of generating default uploader ID for ``MusicManager``
 (see [#2](https://github.com/thebigmunch/google-music/issues/2) for explanation).
+* ``album_art_path`` argument to ``MusicManager.upload`` must
+  now be a relative filename or absolute filepath, not a list.
 
 ### Removed